Revert "Enable test that requires boltkit"
