+++ driver/src/test/java/org/neo4j/driver/internal/RoutingDriverStubTest.java
@@ -18,6 +18,8 @@
  */
 package org.neo4j.driver.internal;
 
+import gherkin.lexer.Tr;
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -43,6 +45,7 @@ import org.neo4j.driver.v1.GraphDatabase;
 import org.neo4j.driver.v1.Record;
 import org.neo4j.driver.v1.Session;
 import org.neo4j.driver.v1.Transaction;
+import org.neo4j.driver.v1.exceptions.ConnectionFailureException;
 import org.neo4j.driver.v1.exceptions.ServiceUnavailableException;
 import org.neo4j.driver.v1.exceptions.SessionExpiredException;
 import org.neo4j.driver.v1.util.Function;
@@ -60,7 +63,8 @@ import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
-public class RoutingDriverBoltKitTest
+@Ignore
+public class RoutingDriverStubTest
 {
     @Rule
     public ExpectedException exception = ExpectedException.none();
