+++ driver/src/test/resources/acquire_endpoints.script
